Move `useFormKeyboardNavigation` to data entry feature
