+++ frontend/src/components/ui/Form/index.ts
@@ -1,3 +1,2 @@
 export * from "./FormLayout";
 export * from "./Form";
-export * from "./useFormKeyboardNavigation";
