+++ frontend/src/features/data_entry/hooks/useDataEntryFormSection.ts
@@ -1,12 +1,12 @@
 import * as React from "react";
 
 import { PollingStationResults } from "@/api";
-import { useFormKeyboardNavigation } from "@/components/ui";
 import { FormSectionId } from "@/types/types";
 
 import { SubmitCurrentFormOptions, TemporaryCache } from "../types/types";
 import { mapValidationResultsToFields } from "../utils/ValidationResults";
 import { useDataEntryContext } from "./useDataEntryContext";
+import { useFormKeyboardNavigation } from "./useFormKeyboardNavigation";
 
 export interface UseDataEntryFormSectionParams<FORM_VALUES> {
   getDefaultFormValues: (results: PollingStationResults, cache?: TemporaryCache | null) => FORM_VALUES;
